Refactor remote git request planning
