+++ docs/DECISIONS.md
@@ -217,3 +217,5 @@
 - Deterministic mock model behavior lives in `MockLLMClient.swift`, not in `Agent.swift`. The mock client can keep local smoke-test and harness heuristics, but the main agent file should stay focused on agent contracts, streaming preview helpers, execution orchestration, and run result types.
 - Streaming action collection and partial assistant-preview parsing live in `AgentActionStreaming.swift`, not in `Agent.swift`. `AgentRunner` and `TrustedRouterLLMClient` should delegate to `AgentActionStreamCollector`, and draft assistant text should go through `AgentActionStreamPreview` so incomplete streaming JSON, raw stream accumulation, and duplicate preview suppression are handled in one tested helper.
 - Sidebar action hierarchy should stay sparse. The visible rail keeps only `New chat` as a primary action; Search, Command palette, Plugins, Automations, Terminal, Browser, Memories, and Activity live in the shared Tools menu driven by `QuillCodeSidebarCommandPresentation.utilityCommandIDs`. This keeps Codex-style navigation calm while preserving quick access and prevents native SwiftUI, static HTML, and Playwright from drifting on sidebar action order.
+- SSH Remote project execution should keep command routing, path validation, and command planning in separate seams. `WorkspaceRemoteProjectToolExecutor` owns only tool routing and SSH execution, `WorkspaceRemoteProjectPath` owns remote path normalization/artifact labels, and `WorkspaceRemoteGitToolRequestPlanner` owns Git/GitHub/worktree command construction plus URL/artifact intent. This keeps remote support testable as SSH, QuillCloud relay, and future remote transports expand.
+- Claude design critique on 2026-06-23 identified the next high-impact UI cleanup: split model selection from approval mode, replace the flat Tools popover with a structured activity surface, teach Auto/Review/Read-only as first-class controls in the empty state, and make approvals a persistent inline queue banner instead of a modal-feeling interruption. Treat those as UX roadmap inputs, not remote-executor scope.
